Don't clobber the underlying c++ executions when aborting, be graceful.
Previously used PyErr_SetInterrupt which would send a keyboard interrupt
back to the controlling python script. However we don't want to abort
in this way - it is preferable to let the execution gracefully bow out.
More information and the actual code that does the bowing out is in
pull request #250.