threadable plasms update.
This reintroduces some level of threading into ecto. It was initially
removed in c2a24a51, but this left some legacy code
wasn't working as expressed in the documentation. These updates
allow parallel execution of plasms in separate threads. It doesn't
quite reach the scope of the original plan for ecto (threading for
free *inside* a plasms computation), but it does cover several useful
scenarios.